Inline and remove manage-ad-free-cookie
